# feat: Add batched GPU inference for 5x performance improvement