@fcakyon 2025-08-09
@bagikazi can you remove example usage py and instead, provide details on how to use batch inference under and md file in docs/ folder?

Also can you move test related utils in the code to a separate test file in tests/ folder?

Also does it support batch inference for all models or only Ultralytics models?

Instead of creating a new independent script, is it possible to implement batch support in models/ultralytics.py or predict.py?